Add files via upload
Research articles containing the terms "poly lactic acid" or PLA or PLLA or PDLA in the title, keyword and abstract were collected using the Scopus database. The used key was TITLE-ABS-KEY ( "poly lactic acid"  OR  pla  OR  plla  OR  pdla)  AND  ( LIMIT-TO ( DOCTYPE ,  "ar" ) )  AND  ( LIMIT-TO ( PUBYEAR ,  2020 )  OR  LIMIT-TO ( PUBYEAR ,  2019 )  OR  LIMIT-TO ( PUBYEAR ,  2018 )  OR  LIMIT-TO ( PUBYEAR ,  2017 )  OR  LIMIT-TO ( PUBYEAR ,  2016 )  OR  LIMIT-TO ( PUBYEAR ,  2015 )  OR  LIMIT-TO ( PUBYEAR ,  2014 )  OR  LIMIT-TO ( PUBYEAR ,  2013 )  OR  LIMIT-TO ( PUBYEAR ,  2012 )  OR  LIMIT-TO ( PUBYEAR ,  2011 )  OR  LIMIT-TO ( PUBYEAR ,  2010 ) ). Articles with publication year of 2021 and 2022 were excluded, as the final number of articles is not consolidated yet. The date of the retrieval was 09/14/2021